+++ CHANGES.md
@@ -49,3 +49,4 @@
 * improve network read error handling
 * added new storage type - git working directory, mirrors git as is, all updates overwrite the content
 * allow compression/encryption in any combination
+* cache uidlist in maildir